+++ src/import.ts
@@ -53,8 +53,7 @@ export function activate(context: ExtensionContext) {
           content
         };
       }));
-
-      await createImporter({
+      const results = await createImporter({
         target,
         ingameDirectory: ingameDirectory.path.replace(/\/$/i, ''),
         result: filesWithContent.reduce((result, item) => {
@@ -63,10 +62,22 @@ export function activate(context: ExtensionContext) {
         }, {}),
         extensionContext: context
       });
+      const successfulItems = results.filter((item) => item.success);
+      const failedItems = results.filter((item) => !item.success);
 
-      vscode.window.showInformationMessage(`${files.length} files got imported to ${ingameDirectory.fsPath}!`, {
-        modal: false
-      });
+      if (successfulItems.length === 0) {
+        vscode.window.showInformationMessage(`No files could get imported! This might be due to a new Grey Hack version or other reasons.`, {
+          modal: false
+        });
+      } else if (failedItems.length > 0) {
+        vscode.window.showInformationMessage(`Import was only partially successful. Only ${successfulItems.length} files got imported to ${ingameDirectory.fsPath}!`, {
+          modal: false
+        });
+      } else {
+        vscode.window.showInformationMessage(`${successfulItems.length} files got imported to ${ingameDirectory.fsPath}!`, {
+          modal: false
+        });
+      }
     } catch (err: any) {
       showCustomErrorMessage(err);
     }